move cronjob into different code fence
